feat(ToolsControl): deport it from index of map controls
